+++ cmp/grid/GridModel.js
@@ -738,11 +738,6 @@ export class GridModel extends HoistModel {
         this.store.clear();
     }
 
-    /** Filter the underlying store. */
-    setFilter(filter) {
-        this.store.setFilter(filter);
-    }
-
     /** @param {Object[]} colConfigs - {@link Column} or {@link ColumnGroup} configs. */
     @action
     setColumns(colConfigs) {