+++ desktop/cmp/leftrightchooser/LeftRightChooserModel.js
@@ -45,8 +45,8 @@ export class LeftRightChooserModel extends HoistModel {
      * @param {function} fn - predicate function for filtering.
      */
     setDisplayFilter(fn) {
-        this.leftModel.setFilter(fn);
-        this.rightModel.setFilter(fn);
+        this.leftModel.store.setFilter(fn);
+        this.rightModel.store.setFilter(fn);
     }
 
     /** Currently 'selected' values on the right hand side. */