+++ CHANGELOG.md
@@ -22,6 +22,8 @@
   `FilterChooserModel.valueSource` and `FilterChooserModel.bind` respectively. Furthermore, both
   configs now support either a `Store` or a cube `View`. This is to provide a common API with the
   new `GridFilterModel` filtering described above.
+* `GridModel.setFilter()` and `DataViewModel.setFilter()` have been removed. Either configure your
+  grid with a `GridFilterModel`, or set the filter on the underlying `Store` instead.
 * `PendingTaskModel` has been replaced by the new `TaskObserver` in `@xh/hoist/core`.
 * The `model` prop on `LoadingIndicator` and `Mask` has been replaced with `bind`.  Provide one or
   more `TaskObserver`s to this prop.